extract npc chatting
