+++ crates/client/src/app/input/mod.rs
@@ -4,3 +4,4 @@ pub mod crafting;
 pub mod hotbar_assignment;
 pub mod inventory;
 pub mod multi_action;
+pub mod npc_interaction;
